CHROMIUM: iwl7000: mac80211: clean up gratuitous upstream diff
There are a number of things that have diverged due to changes in
patches when upstreaming etc. and not coming back through merges.
Clean up the obvious things.

tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: b9f3540371f873b27f82acf227ce79a3fbafb5e2

+++ drivers/net/wireless/iwl7000/mac80211/trace.h
@@ -490,7 +490,7 @@ TRACE_EVENT(drv_vif_cfg_changed,
 
 	TP_printk(
 		LOCAL_PR_FMT  VIF_PR_FMT " changed:%#llx",
-		LOCAL_PR_ARG, VIF_PR_ARG, (unsigned long long)__entry->changed
+		LOCAL_PR_ARG, VIF_PR_ARG, __entry->changed
 	)
 );
 
@@ -564,7 +564,7 @@ TRACE_EVENT(drv_link_info_changed,
 	TP_printk(
 		LOCAL_PR_FMT  VIF_PR_FMT " link_id:%d, changed:%#llx",
 		LOCAL_PR_ARG, VIF_PR_ARG, __entry->link_id,
-		(unsigned long long)__entry->changed
+		__entry->changed
 	)
 );
